Split proto_file up into submodules
