+++ src/proto_node.py
@@ -1,0 +1,24 @@
+from typing import Optional
+
+
+class ProtoNode:
+    @staticmethod
+    def match(proto_source: str) -> Optional["ParsedProtoNode"]:
+        raise NotImplementedError
+
+
+class ParsedProtoNode:
+    def __init__(self, parsed_node: "ProtoNode", remaining_source: str):
+        self.node = parsed_node
+        self.remaining_source = remaining_source
+
+    def __eq__(self, other) -> bool:
+        return (self.node == other.node) and (
+            self.remaining_source == other.remaining_source
+        )
+
+    def __str__(self) -> str:
+        return f"<ParsedProtoNode node={self.node} remaining_source={self.remaining_source} >"
+
+    def __repr__(self) -> str:
+        return str(self)
